add record and listen tabs
