+++ src/views/Journal/components/Navigation/Naivgation.js
@@ -1,4 +1,5 @@
 import React from "react";
+import { Tabs, Tab } from "material-ui";
 
 /* component styles */
 import { styles } from "./styles.scss";
@@ -9,7 +10,14 @@ class Navigation extends React.Component {
   }
 
   render() {
-    return <div className={styles}>nav component</div>;
+    return (
+      <div className={styles}>
+        <Tabs className="tabs">
+          <Tab label="Record" value={"/record"} />
+          <Tab label="Listen" value={"/recordings"} />
+        </Tabs>
+      </div>
+    );
   }
 }
