+++ src/views/Journal/containers/App/JournalView.js
@@ -1,6 +1,5 @@
 import React from "react";
 import JournalHeader from "../JournalHeader/JournalHeader";
-import JournalSubheader from "../JournalSubheader/JournalSubheader";
 import LeftNavBar from "../LeftNavBar/LeftNavBar";
 import RecordView from "../RecordView/RecordView";
 
@@ -12,12 +11,8 @@ import getMuiTheme from "material-ui/styles/getMuiTheme";
 const JournalView = () => (
   <MuiThemeProvider muiTheme={getMuiTheme(darkBaseTheme)}>
     <div>
-      <JournalSubheader />
+      <JournalHeader />
       <div className="container">
-        <h1> (Mic image) </h1>
-        <h1> (Soundwave visual) </h1>
-        <h1> (Recording time) </h1>
-        <h1> Mic photo incoming </h1>
         <LeftNavBar />
       </div>
     </div>
